requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.3.1
distro==1.9.0
ffmpeg-downloader==0.3.0
filelock==3.13.1
fsspec==2024.2.0
gpt4all==2.2.1.post1
greenlet==3.0.3
gTTS==2.5.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
Jinja2==3.1.3
llvmlite==0.42.0
MarkupSafe==2.1.5
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
openai==1.13.3
openai-whisper==20231117
packaging==23.2
platformdirs==4.2.0
psycopg2==2.9.9
pydantic==2.6.3
pydantic_core==2.16.3
pydub==0.25.1
pypiwin32==223
python-dotenv==1.0.1
python-telegram-bot==21.0.1
pyttsx3==2.90
pywin32==306
regex==2023.12.25
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.28
sympy==1.12
tiktoken==0.6.0
torch==2.2.1
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
